+++ CHANGES.rst
@@ -7,7 +7,7 @@ since the branch point at Zope 4.1.2.
 The change log for the previous version, Zope 4, is at
 https://github.com/zopefoundation/Zope/blob/4.x/CHANGES.rst
 
-5.0a2 (unreleased)
+5.0a2 (2020-04-24)
 ------------------
 
 Bug fixes